Refactor aria and element rules
- Add unknown element rule fixes #9
- Fix tests so role attribute is whitespace separated fixes #42
- Refactor standards into aria and elements
- Refactor utils
- Correct list of elements and obsolete elements
