+++ lib/aria/attributes.js
@@ -1,0 +1,54 @@
+/**
+ *  Aria properties
+ */
+
+module.exports = {
+  activedescendant: {},
+  atomic: {},
+  autocomplete: {},
+  busy: {},
+  checked: {},
+  colcount: {},
+  colindex: {},
+  colspan: {},
+  controls: {},
+  current: {},
+  describedby: {},
+  details: {},
+  disabled: {},
+  dropeffect: {},
+  errormessage: {},
+  expanded: {},
+  flowto: {},
+  grabbed: {},
+  haspopup: {},
+  hidden: {},
+  invalid: {},
+  keyshortcuts: {},
+  label: {},
+  labelledby: {},
+  level: {},
+  live: {},
+  modal: {},
+  multiline: {},
+  multiselectable: {},
+  orientation: {},
+  owns: {},
+  placeholder: {},
+  posinset: {},
+  pressed: {},
+  readonly: {},
+  relevant: {},
+  required: {},
+  roledescription: {},
+  rowcount: {},
+  rowindex: {},
+  rowspan: {},
+  selected: {},
+  setsize: {},
+  sort: {},
+  valuemax: {},
+  valuemin: {},
+  valuenow: {},
+  valuetext: {},
+};
