+++ lib/aria/match.js
@@ -1,80 +1,80 @@
 /**
- * Rules for what an element allows
+ * Aria rules for a HTML element
  *
  * https://w3c.github.io/html-aria/
  */
 
-function rule(options = {}) {
-  return Object.assign(
-    {
-      roles: false,
-      aria: true,
-      selector: '*',
-      implicitAria: false,
-    },
-    options,
-    {
-      implicit: [].concat(options.implicit || []),
+const allRoles = require('./roles');
+
+class Rule {
+  constructor({ selector = '*', implicit = [], roles = [], anyRole = false }) {
+    this.selector = selector;
+    this.implicitRoles = [].concat(implicit);
+    this.roles = roles;
+    this.anyRole = anyRole;
+  }
+
+  get allowedRoles() {
+    if (this.anyRole) {
+      return Object.keys(allRoles).filter(name => !this.implicitRoles.includes(name));
     }
-  );
+    return this.roles;
+  }
 }
 
 // Common rules
-const noRoleOrAria = rule({ aria: false });
-const noRole = rule();
-const anyRole = rule({ roles: true });
-
-exports.defaultRule = anyRole;
+// TODO: include aria attribute rules
+const noRoleOrAria = new Rule({});
+const noRole = new Rule({});
+const anyRole = new Rule({ anyRole: true });
 
 // Hash of elements and rules
-exports.rules = {
+const rules = {
   a: [
-    rule({
+    new Rule({
       selector: '[href]',
       implicit: 'link',
       roles: [
         'button', 'checkbox', 'menuitem', 'menuitemcheckbox',
         'menuitemradio', 'radio', 'tab', 'switch', 'treeitem',
       ],
-      implicitAria: true,
     }),
-    rule({
+    new Rule({
       selector: ':not([href])',
-      roles: true,
+      anyRole: true,
     }),
   ],
-  address: rule({
+  address: new Rule({
     implicit: ['contentinfo'],
   }),
   area: [
-    rule({
+    new Rule({
       selector: '[href]',
       implicit: 'link',
-      implicitAria: true,
     }),
   ],
-  article: rule({
+  article: new Rule({
     implicit: 'article',
     roles: ['presentation', 'document', 'application', 'main', 'region'],
   }),
-  aside: rule({
+  aside: new Rule({
     implicit: 'complementary',
     roles: ['note', 'region', 'search'],
   }),
-  audio: rule({
+  audio: new Rule({
     roles: ['application'],
   }),
   base: noRoleOrAria,
-  body: rule({
+  body: new Rule({
     implicit: ['document'],
   }),
   button: [
-    rule({
+    new Rule({
       selector: '[type=menu]',
       implicit: 'button',
       roles: ['link', 'menuitem', 'menuitemcheckbox', 'menuitemradio', 'radio'],
     }),
-    rule({
+    new Rule({
       implicit: 'button',
       roles: [
         'checkbox', 'link', 'menuitem', 'menuitemcheckbox',
@@ -85,171 +85,160 @@ exports.rules = {
   caption: noRole,
   col: noRoleOrAria,
   colgroup: noRoleOrAria,
-  datalist: rule({
+  datalist: new Rule({
     implicit: 'listbox',
-    implicitAria: true,
   }),
-  dd: rule({
+  dd: new Rule({
     implicit: 'definition',
-    implicitAria: true,
   }),
-  details: rule({
+  details: new Rule({
     implicit: 'group',
-    implicitAria: true,
   }),
-  dialog: rule({
+  dialog: new Rule({
     implicit: 'dialog',
     roles: ['alertdialog'],
-    implicitAria: true,
   }),
   div: anyRole,
-  dl: rule({
+  dl: new Rule({
     implicit: 'list',
     roles: ['group', 'presentation'],
   }),
-  dt: rule({
+  dt: new Rule({
     implicit: 'listitem',
-    implicitAria: true,
   }),
-  embed: rule({
+  embed: new Rule({
     roles: ['application', 'document', 'presentation', 'img'],
   }),
-  fieldset: rule({
+  fieldset: new Rule({
     roles: ['group', 'presentation'],
   }),
-  figure: rule({
+  figure: new Rule({
     implicit: 'figure',
     roles: ['group', 'presentation'],
   }),
   footer: [
-    rule({
+    new Rule({
       selector: 'article footer,section footer',
       roles: ['group', 'presentation'],
     }),
-    rule({
+    new Rule({
       implicit: 'contentinfo',
       roles: ['group', 'presentation'],
     }),
   ],
-  form: rule({
+  form: new Rule({
     implicit: 'form',
     roles: ['search', 'presentation'],
   }),
   p: anyRole,
   pre: anyRole,
   blockquote: anyRole,
-  h1: rule({
+  h1: new Rule({
     implicit: 'heading',
     roles: ['tab', 'presentation'],
   }),
-  h2: rule({
+  h2: new Rule({
     implicit: 'heading',
     roles: ['tab', 'presentation'],
   }),
-  h3: rule({
+  h3: new Rule({
     implicit: 'heading',
     roles: ['tab', 'presentation'],
   }),
-  h4: rule({
+  h4: new Rule({
     implicit: 'heading',
     roles: ['tab', 'presentation'],
   }),
-  h5: rule({
+  h5: new Rule({
     implicit: 'heading',
     roles: ['tab', 'presentation'],
   }),
-  h6: rule({
+  h6: new Rule({
     implicit: 'heading',
     roles: ['tab', 'presentation'],
   }),
   head: noRoleOrAria,
   header: [
-    rule({
+    new Rule({
       selector: 'article header,section header',
       roles: ['group', 'presentation'],
     }),
-    rule({
+    new Rule({
       implicit: 'banner',
       roles: ['group', 'presentation'],
     }),
   ],
-  hr: rule({
+  hr: new Rule({
     implicit: 'separator',
     roles: ['presentation'],
-    implicitAria: true,
   }),
   html: noRoleOrAria,
-  iframe: rule({
+  iframe: new Rule({
     roles: ['application', 'document', 'img'],
   }),
   img: [
-    rule({
+    new Rule({
       selector: '[alt=""]',
       roles: ['presentation'],
       aria: false,
     }),
-    rule({
+    new Rule({
       implicit: 'img',
-      roles: true,
+      anyRole: true,
     }),
   ],
   input: [
-    rule({
+    new Rule({
       selector: '[list]:not([type]),[list][type=text],[list][type=search],[list][type=tel],[list][type=url],[list][type=email]',
       implicit: 'combobox',
-      implicitAria: true,
     }),
-    rule({
+    new Rule({
       selector: '[type=button]',
       implicit: 'button',
       roles: ['link', 'menuitem', 'menuitemcheckbox', 'menuitemradio', 'radio', 'switch', 'tab'],
     }),
-    rule({
+    new Rule({
       selector: '[type=image]',
       implicit: 'button',
       roles: ['link', 'menuitem', 'menuitemcheckbox', 'menuitemradio', 'radio', 'switch'],
     }),
-    rule({
+    new Rule({
       selector: '[type=checkbox]',
       implicit: 'checkbox',
       roles: ['button', 'menuitemcheckbox', 'switch'],
     }),
-    rule({
+    new Rule({
       selector: ':not([type]),[type=password],[type=tel],[type=text],[type=url]',
       implicit: 'textbox',
     }),
-    rule({
+    new Rule({
       selector: '[type=email]',
       implicit: 'textbox',
-      implicitAria: true,
     }),
-    rule({
+    new Rule({
       selector: '[type=hidden]',
       aria: false,
     }),
-    rule({
+    new Rule({
       selector: '[type=number]',
       implicit: 'spinbutton',
-      implicitAria: true,
     }),
-    rule({
+    new Rule({
       selector: '[type=radio]',
       implicit: 'radio',
       roles: ['menuitemradio'],
     }),
-    rule({
+    new Rule({
       selector: '[type=range]',
       implicit: 'slider',
     }),
-    rule({
+    new Rule({
       selector: '[type=reset],[type=submit]',
       implicit: 'button',
-      implicitAria: true,
     }),
-    rule({
+    new Rule({
       selector: '[type=search]',
       implicit: 'searchbox',
-      implicitAria: true,
     }),
     noRole,
   ],
@@ -259,7 +248,7 @@ exports.rules = {
   label: noRole,
   legend: noRole,
   li: [
-    rule({
+    new Rule({
       selector: 'ol>li,ul>li',
       implicit: 'listitem',
       roles: [
@@ -269,83 +258,78 @@ exports.rules = {
     }),
   ],
   link: [
-    rule({
+    new Rule({
       selector: '[href]',
       implicit: 'link',
       globalAria: false,
     }),
   ],
-  main: rule({
+  main: new Rule({
     implicit: 'main',
   }),
   map: noRoleOrAria,
-  math: rule({
+  math: new Rule({
     implicit: 'math',
   }),
   menu: [
-    rule({
+    new Rule({
       selector: '[type=toolbar]',
       implicit: 'toolbar',
-      implicitAria: true,
     }),
   ],
   menuitem: [
-    rule({
+    new Rule({
       selector: '[type=command]',
       implicit: 'menuitem',
-      implicitAria: true,
     }),
-    rule({
+    new Rule({
       selector: '[type=checkbox]',
       implicit: 'menuitemcheckbox',
-      implicitAria: true,
     }),
-    rule({
+    new Rule({
       selector: '[type=radio]',
       implicit: 'menuitemradio',
-      implicitAria: true,
     }),
   ],
   meta: noRoleOrAria,
-  meter: rule({
+  meter: new Rule({
     implicit: 'progressbar',
   }),
-  nav: rule({
+  nav: new Rule({
     implicit: 'navigation',
   }),
   noscript: noRoleOrAria,
-  object: rule({
+  object: new Rule({
     roles: ['application', 'document', 'img'],
   }),
-  ol: rule({
+  ol: new Rule({
     implicit: 'list',
     roles: [
       'directory', 'group', 'listbox', 'menu', 'menubar', 'presentation',
       'radiogroup', 'tablist', 'toolbar', 'tree',
     ],
   }),
-  optgroup: rule({
+  optgroup: new Rule({
     implicit: 'group',
   }),
   option: [
-    rule({
+    new Rule({
       selector: 'select>option,select>optgroup>option,datalist>option',
       implicit: 'option',
     }),
     noRoleOrAria,
   ],
-  output: rule({
+  output: new Rule({
     implicit: 'status',
-    roles: true,
+    anyRole: true,
   }),
   param: noRoleOrAria,
   picture: noRoleOrAria,
-  progress: rule({
+  progress: new Rule({
     implicit: 'progressbar',
-    implicitAria: true,
   }),
   script: noRoleOrAria,
-  section: rule({
+  section: new Rule({
     implicit: 'region',
     roles: [
       'alert', 'alertdialog', 'application', 'banner', 'complementary',
@@ -353,43 +337,42 @@ exports.rules = {
       'navigation', 'search', 'status',
     ],
   }),
-  select: rule({
+  select: new Rule({
     implicit: 'listbox',
   }),
   source: noRoleOrAria,
   span: anyRole,
   style: noRoleOrAria,
-  svg: rule({
+  svg: new Rule({
     roles: ['application', 'document', 'img'],
   }),
-  summary: rule({
+  summary: new Rule({
     implicit: 'button',
-    implicitAria: true,
   }),
-  table: rule({
+  table: new Rule({
     implicit: 'table',
-    roles: true,
+    anyRole: true,
   }),
   template: noRoleOrAria,
-  textarea: rule({
+  textarea: new Rule({
     implicit: 'textbox',
   }),
-  tbody: rule({
+  tbody: new Rule({
     implicit: 'rowgroup',
-    roles: true,
+    anyRole: true,
   }),
-  thead: rule({
+  thead: new Rule({
     implicit: 'rowgroup',
-    roles: true,
+    anyRole: true,
   }),
-  tfoot: rule({
+  tfoot: new Rule({
     implicit: 'rowgroup',
-    roles: true,
+    anyRole: true,
   }),
   title: noRoleOrAria,
-  td: rule({
+  td: new Rule({
     implicit: 'cell',
-    roles: true,
+    anyRole: true,
   }),
   em: anyRole,
   strong: anyRole,
@@ -417,23 +400,32 @@ exports.rules = {
   bdo: anyRole,
   br: anyRole,
   wbr: anyRole,
-  th: rule({
+  th: new Rule({
     implicit: ['columnheader', 'rowheader'],
-    roles: true,
+    anyRole: true,
   }),
-  tr: rule({
+  tr: new Rule({
     implicit: 'row',
-    roles: true,
+    anyRole: true,
   }),
   track: noRoleOrAria,
-  ul: rule({
+  ul: new Rule({
     implicit: 'list',
     roles: [
       'directory', 'group', 'listbox', 'menu', 'menubar',
       'tablist', 'toolbar', 'tree', 'presentation',
     ],
   }),
-  video: rule({
+  video: new Rule({
     roles: ['application'],
   }),
+};
+
+module.exports = function match(el) {
+  const name = el.nodeName.toLowerCase();
+  let found = rules[name];
+  if (Array.isArray(found)) {
+    found = found.find(item => item.selector === '*' || el.matches(item.selector));
+  }
+  return found || anyRole;
 };
