+++ lib/aria/roles.js
@@ -4,98 +4,137 @@
  * https://w3c.github.io/html-aria/
  */
 
-// Aria properties that can be used on any HTML element
-const globalAria = [
-  'atomic', 'busy', 'controls', 'current', 'describedby', 'details',
-  'disabled', 'dropeffect', 'errormessage', 'flowto', 'grabbed',
-  'haspopup', 'hidden', 'invalid', 'keyshortcuts', 'label',
-  'labelledby', 'live', 'owns', 'relevant', 'roledescription',
-];
-
-
-// All roles and their allowed aria properties
-const expanded = {
-  allowed: ['expanded'],
-};
-
-const roles = {
-  alert: expanded,
+module.exports = {
+  alert: {
+    allowed: ['expanded'],
+  },
   alertdialog: {
     allowed: ['expanded', 'modal'],
   },
   application: {
     allowed: ['activedescendant'],
   },
-  article: expanded,
-  banner: expanded,
+  article: {
+    allowed: ['expanded'],
+  },
+  banner: {
+    allowed: ['expanded'],
+  },
   button: {
     allowed: ['expanded', 'pressed'],
+    nameFromContent: true,
   },
-  cell: {},
+  cell: {
+    nameFromContent: true,
+  },
   checkbox: {
     required: ['checked'],
+    nameFromContent: true,
   },
   columnheader: {
     allowed: ['sort', 'readonly', 'required', 'selected', 'expanded'],
+    nameFromContent: true,
   },
   combobox: {
     required: ['expanded'],
     allowed: ['autocomplete', 'required', 'activedescendant'],
   },
-  complementary: expanded,
-  contentinfo: expanded,
-  definition: expanded,
-  dialog: expanded,
-  directory: expanded,
-  document: expanded,
+  complementary: {
+    allowed: ['expanded'],
+  },
+  contentinfo: {
+    allowed: ['expanded'],
+  },
+  definition: {
+    allowed: ['expanded'],
+  },
+  dialog: {
+    allowed: ['expanded'],
+  },
+  directory: {
+    allowed: ['expanded'],
+  },
+  document: {
+    allowed: ['expanded'],
+  },
   feed: {
     allowed: ['setsize', 'expanded'],
   },
-  figure: expanded,
-  form: expanded,
+  figure: {
+    allowed: ['expanded'],
+  },
+  form: {
+    allowed: ['expanded'],
+  },
   grid: {
     allowed: ['level', 'multiselectable', 'readonly', 'activedescendant', 'expanded'],
   },
   gridcell: {
     allowed: ['readonly', 'required', 'selected', 'expanded'],
+    nameFromContent: true,
   },
   group: {
     allowed: ['activedescendant', 'expanded'],
   },
   heading: {
     allowed: ['level', 'expanded'],
+    nameFromContent: true,
   },
-  img: expanded,
-  link: expanded,
-  list: expanded,
+  img: {
+    allowed: ['expanded'],
+  },
+  link: {
+    allowed: ['expanded'],
+    nameFromContent: true,
+  },
+  list: {
+    allowed: ['expanded'],
+  },
   listbox: {
     allowed: ['multiselectable', 'required', 'expanded', 'activedescendant', 'expanded'],
   },
   listitem: {
     allowed: ['level', 'posinset', 'setsize', 'expanded'],
   },
-  log: expanded,
-  main: expanded,
-  marquee: expanded,
-  math: expanded,
+  log: {
+    allowed: ['expanded'],
+  },
+  main: {
+    allowed: ['expanded'],
+  },
+  marquee: {
+    allowed: ['expanded'],
+  },
+  math: {
+    allowed: ['expanded'],
+  },
   menu: {
     allowed: ['activedescendant', 'expanded'],
   },
   menubar: {
     allowed: ['activedescendant'],
   },
-  menuitem: {},
+  menuitem: {
+    nameFromContent: true,
+  },
   menuitemcheckbox: {
     required: ['checked'],
+    nameFromContent: true,
   },
   menuitemradio: {
     required: ['checked'],
     allowed: ['posinset', 'selected', 'setsize'],
+    nameFromContent: true,
   },
-  navigation: expanded,
-  note: expanded,
+  navigation: {
+    allowed: ['expanded'],
+  },
+  note: {
+    allowed: ['expanded'],
+  },
   option: {
     allowed: ['checked', 'posinset', 'selected', 'setsize'],
+    nameFromContent: true,
   },
   presentation: {},
   progressbar: {
@@ -104,22 +143,28 @@ const roles = {
   radio: {
     required: ['checked'],
     allowed: ['posinset', 'selected', 'setsize'],
+    nameFromContent: true,
   },
   radiogroup: {
     allowed: ['required', 'activedescendant', 'expanded'],
   },
-  region: expanded,
+  region: {
+    allowed: ['expanded'],
+  },
   row: {
     allowed: [
       'colindex', 'level', 'rowindex', 'selected', 'level', 'selected',
       'activedescendant', 'expanded',
     ],
+    nameFromContent: true,
   },
   rowgroup: {
     allowed: ['activedescendant', 'expanded'],
+    nameFromContent: true,
   },
   rowheader: {
     allowed: ['sort', 'readonly', 'required', 'selected', 'expanded'],
+    nameFromContent: true,
   },
   scrollbar: {
     required: ['controls', 'orientation', 'valuemax', 'valuemin', 'valuenow'],
@@ -142,12 +187,16 @@ const roles = {
     required: ['valuemax', 'valuemin', 'valuenow'],
     allowed: ['required', 'valuetext'],
   },
-  status: expanded,
+  status: {
+    allowed: ['expanded'],
+  },
   switch: {
     required: ['checked'],
+    nameFromContent: true,
   },
   tab: {
     allowed: ['selected', 'expanded'],
+    nameFromContent: true,
   },
   table: {
     allowed: ['colcount', 'rowcount'],
@@ -155,35 +204,33 @@ const roles = {
   tablist: {
     allowed: ['level', 'activedescendant', 'expanded'],
   },
-  tabpanel: expanded,
-  term: expanded,
+  tabpanel: {
+    allowed: ['expanded'],
+  },
+  term: {
+    allowed: ['expanded'],
+  },
   textbox: {
     allowed: ['activedescendant', 'autocomplete', 'multiline', 'placeholder', 'readonly', 'required'],
   },
-  timer: expanded,
+  timer: {
+    allowed: ['expanded'],
+  },
   toolbar: {
     allowed: ['activedescendant', 'expanded'],
+    nameFromContent: true,
   },
-  tooltip: expanded,
+  tooltip: {
+    allowed: ['expanded'],
+  },
   tree: {
     allowed: ['multiselectable', 'required', 'activedescendant', 'expanded'],
+    nameFromContent: true,
   },
-  treegrid: ['level', 'multiselecteable', 'readonly', 'activedescendant', 'expanded', 'required'],
-  treeitem: ['level', 'posinset', 'setsize', 'expanded', 'checked', 'selected'],
+  treegrid: {
+    allowed: ['level', 'multiselecteable', 'readonly', 'activedescendant', 'expanded', 'required'],
+  },
+  treeitem: {
+    allowed: ['level', 'posinset', 'setsize', 'expanded', 'checked', 'selected'],
+  },
 };
-
-const bool = ['true', 'false'];
-const tri = ['true', 'false', 'mixed'];
-
-// Allowed aria property values
-const properties = {
-  activedescendant: 'id',
-  atomic: bool,
-  autocomplete: ['inline', 'list', 'both', 'none'],
-  busy: bool,
-  checked: tri,
-};
-
-exports.globalAria = globalAria;
-exports.roles = roles;
-exports.properties = properties;
