+++ lib/elements.js
@@ -1,5 +1,11 @@
+/**
+ * Data for HTML elements.  Based on
+ * - https://www.w3.org/TR/html52/
+ * - https://w3c.github.io/html-aria/
+ */
+
 const obsolete = { obsolete: true };
-exports.elements = {
+module.exports = {
   a: {},
   abbr: {},
   acronym: obsolete,
@@ -28,8 +34,7 @@ exports.elements = {
   code: {},
   col: {},
   colgroup: {},
-  command: {},
-  content: {},
+  command: obsolete,
   data: {},
   datalist: {},
   dd: {},
@@ -41,7 +46,6 @@ exports.elements = {
   div: {},
   dl: {},
   dt: {},
-  element: {},
   em: {},
   embed: {},
   fieldset: {},
@@ -60,18 +64,18 @@ exports.elements = {
   h6: {},
   head: {},
   header: {},
-  hgroup: {},
+  hgroup: obsolete,
   hr: {},
   html: {},
   i: {},
   iframe: {},
-  image: {},
+  image: obsolete,
   img: {},
   input: {},
   ins: {},
   isindex: obsolete,
   kbd: {},
-  keygen: {},
+  keygen: obsolete,
   label: {},
   legend: {},
   li: {},
@@ -104,17 +108,16 @@ exports.elements = {
   pre: {},
   progress: {},
   q: {},
-  rb: obsolete,
+  rb: {},
   rp: {},
   rt: {},
-  rtc: obsolete,
+  rtc: {},
   ruby: {},
   s: {},
   samp: {},
   script: {},
   section: {},
   select: {},
-  shadow: {},
   small: {},
   source: {},
   spacer: obsolete,
