+++ lib/linter.js
@@ -3,7 +3,8 @@ const Logger = require('./logger');
 const rules = require('./rules');
 const utils = require('./utils');
 const version = require('./version');
-const standards = require('./standards');
+const aria = require('./aria');
+const elements = require('./elements');
 
 const Linter = module.exports = class AccessibilityLinter extends Runner {
   constructor(options) {
@@ -36,4 +37,6 @@ const Linter = module.exports = class AccessibilityLinter extends Runner {
 Linter.Logger = Logger;
 Linter.rules = rules;
 Linter.version = version;
-Linter.standards = standards;
+Linter.aria = aria;
+Linter.elements = elements;
+Linter.utils = utils;
