+++ lib/utils/index.js
@@ -1,4 +1,10 @@
 /**
+ * Utils for working with the DOM
+ */
+
+const hidden = require('./hidden');
+
+/**
  * Find DOM nodes from a selector.  The found node can include the supplied context
  * @param {String|NodeList} selector
  * @param {HTMLElement} [context]
@@ -19,6 +25,8 @@ exports.$ = function $(selector, context) {
 exports.cssEscape = function cssEscape(name) {
   return name.replace(/["\\]/g, '\\$&');
 };
+
+exports.hidden = hidden;
 
 /**
  * Observe for child list mutations
