+++ rules/aria/roles/rule.js
@@ -1,19 +1,36 @@
 ({
   message(el) {
-    const rule = standards.aria.match(el);
-    const role = el.getAttribute('role');
-    if (rule.implicitRoles.includes(role)) {
-      return `role "${role}" is implicit for this element and not allowed`;
+    return this.check(el);
+  },
+  check(el) {
+    const role = el.getAttribute('role').trim();
+    if (!role) {
+      return 'role attribute should not be empty';
     }
-    if (!standards.aria.roles.includes(role)) {
-      return `role "${role}" is not a known role`;
-    }
-    return `role "${role}" is not allowed for this element`;
+    let error;
+    const rule = aria.match(el);
+    role.split(/\s+/).some((name) => {
+      if (!aria.roles[name]) {
+        error = `"${name}" is not a known role`;
+        return true;
+      }
+
+      if (rule.implicitRoles.includes(name)) {
+        error = `role "${name}" is implicit for this element and should not be specified`;
+        return true;
+      }
+
+      if (!rule.allowedRoles.includes(name)) {
+        error = `role "${name}" is not allowed for this element`;
+      }
+
+      return false;
+    });
+
+    return error;
   },
   selector: '[role]',
   filter(el) {
-    const rule = standards.aria.match(el);
-    const role = el.getAttribute('role');
-    return rule.allowedRoles.includes(role);
+    return !this.check(el);
   },
 });
