+++ rules/elements/obsolete/rule.js
@@ -1,0 +1,4 @@
+({
+  message: 'do not use obsolete elements',
+  selector: Object.keys(elements).filter(el => elements[el].obsolete).join(','),
+});
