+++ rules/elements/obsolete/spec.js
