+++ rules/elements/unknown/rule.js
@@ -1,0 +1,4 @@
+({
+  message: 'unknown element',
+  selector: Object.keys(elements).map(name => `:not(${name})`).join(''),
+});
