--- rules/obsolete/elements/rule.js
@@ -1,4 +1,0 @@
-({
-  message: 'do not use obsolete elements',
-  selector: Array.from(standards.elements.obsolete),
-});
