+++ test/specs/elements.js
@@ -1,0 +1,163 @@
+describe('elements', () => {
+  const elements = AccessibilityLinter.elements;
+
+  it('is a property of AccessibilityLinter', () => {
+    expect(elements).toExist();
+  });
+
+  const obsolete = [
+    'applet',
+    'acronym',
+    'bgsound',
+    'command',
+    'dir',
+    'frame',
+    'frameset',
+    'hgroup',
+    'image',
+    'noframes',
+    'isindex',
+    'keygen',
+    'listing',
+    'nextid',
+    'noembed',
+    'plaintext',
+    'strike',
+    'xmp',
+    'basefont',
+    'big',
+    'blink',
+    'center',
+    'font',
+    'marquee',
+    'multicol',
+    'nobr',
+    'spacer',
+    'tt',
+  ].sort();
+
+  const allElements = [
+    'a',
+    'abbr',
+    'address',
+    'area',
+    'article',
+    'aside',
+    'audio',
+    'b',
+    'base',
+    'bdi',
+    'bdo',
+    'blockquote',
+    'body',
+    'br',
+    'button',
+    'canvas',
+    'caption',
+    'cite',
+    'code',
+    'col',
+    'colgroup',
+    'data',
+    'datalist',
+    'dd',
+    'del',
+    'details',
+    'dfn',
+    'dialog',
+    'div',
+    'dl',
+    'dt',
+    'em',
+    'embed',
+    'fieldset',
+    'figcaption',
+    'figure',
+    'footer',
+    'form',
+    'h1',
+    'h2',
+    'h3',
+    'h4',
+    'h5',
+    'h6',
+    'head',
+    'header',
+    'hr',
+    'html',
+    'i',
+    'iframe',
+    'img',
+    'input',
+    'ins',
+    'kbd',
+    'label',
+    'legend',
+    'li',
+    'link',
+    'main',
+    'map',
+    'mark',
+    'menu',
+    'menuitem',
+    'meta',
+    'meter',
+    'nav',
+    'noscript',
+    'object',
+    'ol',
+    'optgroup',
+    'option',
+    'output',
+    'p',
+    'param',
+    'picture',
+    'pre',
+    'progress',
+    'q',
+    'rb',
+    'rp',
+    'rt',
+    'rtc',
+    'ruby',
+    's',
+    'samp',
+    'script',
+    'section',
+    'select',
+    'small',
+    'source',
+    'span',
+    'strong',
+    'style',
+    'sub',
+    'summary',
+    'sup',
+    'table',
+    'tbody',
+    'td',
+    'template',
+    'textarea',
+    'tfoot',
+    'th',
+    'thead',
+    'time',
+    'title',
+    'tr',
+    'track',
+    'u',
+    'ul',
+    'var',
+    'video',
+    'wbr',
+  ];
+
+  it('contains all elements', () => {
+    expect(Object.keys(elements).sort()).toEqualArray(allElements.concat(obsolete).sort());
+  });
+
+  it('contains all obsolete elements', () => {
+    expect(Object.keys(elements).filter(name => elements[name].obsolete).sort())
+      .toEqualArray(obsolete);
+  });
+});
