+++ test/test.htm
@@ -11,11 +11,15 @@
   <script src="../node_modules/expect/umd/expect.js"></script>
   <script>mocha.setup('bdd');</script>
   <script src="../build/umd.js"></script>
-  <script src="rule-specs.js"></script>
   <script src="support.js"></script>
   <script src="utils.js"></script>
-  <script src="spec.js"></script>
-  <script src="standards-spec.js"></script>
+  <script src="specs/_rules.js"></script>
+  <script src="specs/linter.js"></script>
+  <script src="specs/logger.js"></script>
+  <script src="specs/utils.js"></script>
+  <script src="specs/aria.js"></script>
+  <script src="specs/elements.js"></script>
+  <script src="specs/rules.js"></script>
   <script>
     mocha.checkLeaks();
     mocha.run();
